Sonar code smells
