+++ CoreFindView/src/au/gov/asd/tac/constellation/views/find2/plugins/BasicFindPlugin.java
@@ -19,7 +19,6 @@ import au.gov.asd.tac.constellation.graph.Attribute;
 import au.gov.asd.tac.constellation.graph.Graph;
 import au.gov.asd.tac.constellation.graph.GraphElementType;
 import au.gov.asd.tac.constellation.graph.GraphWriteMethods;
-import au.gov.asd.tac.constellation.graph.interaction.gui.VisualGraphTopComponent;
 import au.gov.asd.tac.constellation.graph.schema.visual.concept.VisualConcept;
 import au.gov.asd.tac.constellation.plugins.PluginInfo;
 import au.gov.asd.tac.constellation.plugins.PluginInteraction;
@@ -30,14 +29,11 @@ import au.gov.asd.tac.constellation.views.find2.state.FindViewConcept;
 import au.gov.asd.tac.constellation.views.find2.utilities.BasicFindReplaceParameters;
 import au.gov.asd.tac.constellation.views.find2.utilities.FindResult;
 import au.gov.asd.tac.constellation.views.find2.utilities.FindResultsList;
-import java.awt.EventQueue;
+import au.gov.asd.tac.constellation.views.find2.utilities.FindViewUtilities;
 import java.util.List;
-import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import org.openide.windows.TopComponent;
-import org.openide.windows.WindowManager;
 
 /**
  * This class does the actual action of finding.
@@ -254,19 +250,7 @@ public class BasicFindPlugin extends SimpleEditPlugin {
 
         // Swap to view the graph where the element is selected
         if (searchAllGraphs) {
-            final Set<TopComponent> topComponents = WindowManager.getDefault().getRegistry().getOpened();
-            if (topComponents != null) {
-                for (final TopComponent component : topComponents) {
-                    if ((component instanceof VisualGraphTopComponent) && ((VisualGraphTopComponent) component).getGraphNode().getGraph().getId().equals(graph.getId())) {
-                        EventQueue.invokeLater(new Runnable() {
-                            @Override
-                            public void run() {
-                                ((VisualGraphTopComponent) component).requestActive();
-                            }
-                        });
-                    }
-                }
-            }
+            FindViewUtilities.searchAllGraphs(graph);
         }
 
         //If no results are found, set the meta attribute to null
