+++ CoreFindView/src/au/gov/asd/tac/constellation/views/find2/plugins/ReplacePlugin.java
@@ -18,7 +18,6 @@ package au.gov.asd.tac.constellation.views.find2.plugins;
 import au.gov.asd.tac.constellation.graph.Attribute;
 import au.gov.asd.tac.constellation.graph.GraphElementType;
 import au.gov.asd.tac.constellation.graph.GraphWriteMethods;
-import au.gov.asd.tac.constellation.graph.interaction.gui.VisualGraphTopComponent;
 import au.gov.asd.tac.constellation.graph.schema.visual.concept.VisualConcept;
 import au.gov.asd.tac.constellation.plugins.PluginInfo;
 import au.gov.asd.tac.constellation.plugins.PluginInteraction;
@@ -26,13 +25,10 @@ import au.gov.asd.tac.constellation.plugins.PluginType;
 import au.gov.asd.tac.constellation.plugins.parameters.PluginParameters;
 import au.gov.asd.tac.constellation.plugins.templates.SimpleEditPlugin;
 import au.gov.asd.tac.constellation.views.find2.utilities.BasicFindReplaceParameters;
-import java.awt.EventQueue;
+import au.gov.asd.tac.constellation.views.find2.utilities.FindViewUtilities;
 import java.util.List;
-import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import org.openide.windows.TopComponent;
-import org.openide.windows.WindowManager;
 
 /**
  * Finds String based attributes within the graph and replaces its value.
@@ -134,19 +130,7 @@ public class ReplacePlugin extends SimpleEditPlugin {
 
         // Swap to view the graph where the element is selected 
         if (searchAllGraphs) {
-            final Set<TopComponent> topComponents = WindowManager.getDefault().getRegistry().getOpened();
-            if (topComponents != null) {
-                for (final TopComponent component : topComponents) {
-                    if ((component instanceof VisualGraphTopComponent) && ((VisualGraphTopComponent) component).getGraphNode().getGraph().getId().equals(graph.getId())) {
-                        EventQueue.invokeLater(new Runnable() {
-                            @Override
-                            public void run() {
-                                ((VisualGraphTopComponent) component).requestActive();
-                            }
-                        });
-                    }
-                }
-            }
+            FindViewUtilities.searchAllGraphs(graph);
         }
     }
